requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
pydantic-settings
sentence-transformers
faiss-cpu
openai
tiktoken
python-multipart
streamlit